requirements.txt
mpp-solar>=0.16.0
paho-mqtt>=1.6.0
pyyaml>=6.0